Added preference to enable debug info stripping Added some code by Nick Shanks Some 64-bit fixes

+++ Monolingual/ISTableView.m
@@ -12,7 +12,7 @@
 
 - (void) keyDown:(NSEvent *)theEvent
 {
-	int row;
+	NSInteger row;
 	unichar key = [[theEvent charactersIgnoringModifiers] characterAtIndex:0];
 
 	switch (key) {
